+++ src/main/java/wooteco/subway/member/controller/MemberController.java
@@ -1,12 +1,12 @@
-package wooteco.subway.member.ui;
+package wooteco.subway.member.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import wooteco.subway.auth.config.AuthenticationPrincipal;
-import wooteco.subway.member.application.MemberService;
+import wooteco.subway.member.domain.AuthenticationPrincipal;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
+import wooteco.subway.member.service.MemberService;
 
 import java.net.URI;
