+++ src/main/java/wooteco/subway/member/controller/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.dto;
+package wooteco.subway.member.controller.response;
 
 public class TokenResponse {
     private String accessToken;
