+++ src/main/java/wooteco/subway/member/domain/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.config;
+package wooteco.subway.member.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
