+++ src/main/java/wooteco/subway/member/infra/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package wooteco.subway.auth.infrastructure;
+package wooteco.subway.member.infra;
 
 import io.jsonwebtoken.*;
 import org.springframework.beans.factory.annotation.Value;
