+++ src/main/java/wooteco/subway/member/method/AuthenticationPrincipalConfig.java
@@ -1,11 +1,10 @@
-package wooteco.subway.auth.config;
+package wooteco.subway.member.method;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import wooteco.subway.auth.application.AuthService;
-import wooteco.subway.auth.ui.AuthenticationPrincipalArgumentResolver;
+import wooteco.subway.member.service.AuthService;
 
 import java.util.List;
